Create Hi-Lo Game
Introduction
The objective of this assignment is to master the use of selective and nested repetitive control flow. You will write a program that plays a game called Hi-Lo (開口中). In the game, a secret number between 1 and 100 (inclusive) is generated randomly. Multiple players take turns to make guesses and will be responded with a HI or LO message. HI means the guess is larger than the secret number and LO means the guess is smaller than the secret number. With more and more guesses, the possible range of the secret number becomes narrower and narrower. The goal of the players is to avoid revealing the secret number. Multiple games will be played, each with a different secret number. The players will hope to reveal the secret number the fewest times.