Create sos
Introduction
The objective of this assignment is to practice (1) defining functions (being a callee), (2) calling functions (being a caller), and (3) representing special kind of data. You will implement a two-player paper-and-pencil game called SOS. Two players take turns to mark the spaces in a 3×4 grid with either symbols O or S. Players can use different symbols in different turns. The player who first creates the sequence S-O-S horizontally ―, vertically |, or diagonally ⟍⟋ in the grid wins the game. When the grid is full without any S-O-S sequences, the game is a draw. Error! Reference source not found. shows an example grid. If an S is put into the empty space in the second row, an S-O-S sequence will be formed.